requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-celery-beat==1.5.0
django-guardian==2.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.4
djangorestframework-guardian==0.1.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
